+++ alonet/deformable_detr/train.py
@@ -89,8 +89,6 @@ class LitDeformableDetr(LitDetr):
         loss_label_weight=1,
         loss_boxes_weight=5,
         loss_giou_weight=2,
-        loss_dice_weight=0,
-        loss_focal_weight=0,
         losses=["labels", "boxes"],
         aux_loss_stage=6,
         eos_coef=0.1,
@@ -124,8 +122,6 @@ class LitDeformableDetr(LitDetr):
             loss_label_weight=loss_label_weight,
             loss_boxes_weight=loss_boxes_weight,
             loss_giou_weight=loss_giou_weight,
-            loss_dice_weight=loss_dice_weight,
-            loss_focal_weight=loss_focal_weight,
             losses=losses,
             aux_loss_stage=aux_loss_stage,
             eos_coef=eos_coef,
