+++ alonet/detr/train.py
@@ -197,8 +197,6 @@ class LitDetr(pl.LightningModule):
         loss_ce_weight=1,
         loss_boxes_weight=5,
         loss_giou_weight=2,
-        loss_dice_weight=0,
-        loss_focal_weight=0,
         eos_coef=0.1,
         losses=["labels", "boxes"],
         aux_loss_stage=6,
@@ -208,8 +206,6 @@ class LitDetr(pl.LightningModule):
             loss_ce_weight=loss_ce_weight,
             loss_boxes_weight=loss_boxes_weight,
             loss_giou_weight=loss_giou_weight,
-            loss_dice_weight=loss_dice_weight,
-            loss_focal_weight=loss_focal_weight,
             eos_coef=eos_coef,
             losses=losses,
             aux_loss_stage=aux_loss_stage,
