+++ alonet/detr_panoptic/__init__.py
@@ -1,3 +1,4 @@
 from .detr_panoptic import PanopticHead
+from .criterion import PanopticCriterion
 from .callbacks import PanopticObjectDetectorCallback
 from .train import LitPanopticDetr
